{{ cookiecutter.repo_name }}/requirements.txt
# local package
-e .

# analysis
matplotlib
numpy
pandas
scikit-learn
seaborn
lightgbm
tensorflow
torch

# work
ipython
jupyter
japanize-matplotlib

# pypi
wheel
twine

# external requirements
PyYAML
Sphinx
attrdict
click
coverage
flake8
ipdb
python-dotenv>=0.5.1
sphinx_rtd_theme
tqdm
{% if cookiecutter.python_interpreter != 'python3' %}

# backwards compatibility
pathlib2
{% endif %}
